+++ stdin/stdin.go
@@ -2,20 +2,18 @@
 // Copyright (c) 2014 Fatih Cetinkaya (http://github.com/cmfatih/yapi)
 // For the full copyright and license information, please view the LICENSE.txt file.
 
-// This file contains stdin related implementations.
+// Package stdin provides stdin related functions.
 //
 // References:
 //  `ModeNamedPipe`         : http://golang.org/pkg/os/#FileMode
 //  `http.DetectContentType`: http://golang.org/src/pkg/net/http/sniff.go
 //  File signatures         : http://www.garykessler.net/library/file_sigs.html
 //  Media Types             : http://www.iana.org/assignments/media-types/media-types.xhtml
-
-package pipe
+package stdin
 
 import (
 	"bufio"
-	"errors"
-	//"log"
+	"io"
 	"net/http"
 	"os"
 	"runtime"
@@ -23,13 +21,27 @@ import (
 )
 
 var (
-	contentType string
+	stdinReader    = bufio.NewReader(os.Stdin)
+	stdinHasPipe   bool
+	stdinErr       error
+	contentSize    int64
+	contentType    string
+	contentTypeErr error
 )
 
-func init() {
+// StdinReader returns the stdin reader
+func StdinReader() *bufio.Reader {
+	return stdinReader
+}
 
-	// Init vars
-	contentType, _ = contentTypeDetect()
+// StdinHasPipe returns whether is there a stream on stdin or not
+func StdinHasPipe() bool {
+	return stdinHasPipe
+}
+
+// StdinErr returns the stdin error if any
+func StdinErr() error {
+	return stdinErr
 }
 
 // ContentType returns the content type
@@ -37,12 +49,16 @@ func ContentType() string {
 	return contentType
 }
 
-// contentTypeDetect detects and returns the content type
-func contentTypeDetect() (string, error) {
+// contentTypeError returns the content type error
+func ContentTypeError() error {
+	return contentTypeErr
+}
+
+func init() {
 
 	// TODO: Add more file signatures
 
-	// *nix
+	// Posix
 	//
 	// None    : fi.Size(): 0  / fi.Mode(): Dcrw--w----
 	// |       : fi.Size(): 0  / fi.Mode(): prw-------
@@ -62,31 +78,41 @@ func contentTypeDetect() (string, error) {
 	if runtime.GOOS == "windows" {
 		if err != nil {
 			// no pipe
-			return "", nil
+			// if there is no pipe than os.Stdin.Stat() returns error on Windows
+			stdinHasPipe = false
+		} else {
+			stdinHasPipe = true
 		}
 	} else {
 		if err != nil {
-			return "", errors.New("failed to determine content type: " + err.Error())
+			stdinErr = err
+			stdinHasPipe = false
 		} else if fi.Mode()&os.ModeNamedPipe == 0 && fi.Size() == 0 {
 			// no pipe
-			return "", nil
+			stdinHasPipe = false
+		} else {
+			stdinHasPipe = true
 		}
 	}
 
-	//log.Printf("name: %s - size: %d - mode: %s\n", os.Stdin.Name(), fi.Size(), fi.Mode())
+	if stdinHasPipe == true {
+		contentSize = fi.Size()
+		buf, err := stdinReader.Peek(512)
+		//print("buf: ", string(buf[0:10]), "\n") // for debug
 
-	// Init vars
-	ct := ""
-	rd := bufio.NewReader(os.Stdin)
-	buf, err := rd.Peek(512)
-	dct := http.DetectContentType(buf)
+		if err != nil && err != io.EOF {
+			contentTypeErr = err
+		} else {
+			dct := http.DetectContentType(buf)
 
-	if strings.HasPrefix(dct, "text/") == true {
-		dctf := strings.Split(dct, ";")
-		ct = dctf[0]
-	} else {
-		ct = dct
+			if strings.HasPrefix(dct, "text/") == true {
+				dctf := strings.Split(dct, ";")
+				contentType = dctf[0]
+			} else {
+				contentType = dct
+			}
+		}
 	}
 
-	return ct, nil
+	//print("shp|ct|size: ", stdinHasPipe, "|", contentType, "|", contentSize, "\n") // for debug
 }